Max and I completed the full view.
We added one test to test the contents of the full view page, and added a
assertion to the test_index on line 17 of blog.py.

I added a function to the blog.py to use so even a unlogged in user can still
view each post.

Also, I added a full.html page which is the full blog post view.

Overall, The view extra feature was created.